Implement RoundRect
It seems that Graffle implements RoundRect as the intersection of a rectangle width x height and a disc of diameter width. I did not find an easy way to generate such shapes in SVG.

The present implementation makes use of rx and ry attributes of rect, to mimic the shape described above. To achieve this, rx and ry which represent the radius of an ellipse rounding the corners of the rectangle are chosen to be:

* ry = height/2 : smoothing the circular shape
* rx to match the point at which the disc would intersect the rectangle. Doing some simple pythagor to compute the position.

This closes #3